Bugfix for 38558
Fixed a bug where, when entering the Avatar Menu, the turn-based battle pointer would denullify with garbage data, making Olympus Coliseum achievements believe that a turn-based battle had begun. Leaving the Avatar Menu with a blank save file as the second slot would cause the enemy data memory to flicker, making the achievement believe that the player had defeated the Ice Titan when, in fact, they did not.