Fixed an issue related to the previous one where endgame achievements could pop under certain situations
It used to be the case that, when the following occurred:

* Game was completed as any character on low difficulties
* A new game was started on Critical Mode

...endgame achievements for the character that was just completed would pop when the new game started. This was due to hit targets not resetting as expected when the player returned to the title screen.

A workaround was employed where the achievement will wait one frame longer than the initial 60 frames if it isn't supposed to pop, and if it does, this means that the player will certainly never earn the achievement at that time. One frame later, it will reset.

This could have the unintended side effect of making the achievement pop up sooner than expected once the player completes another character's story, but the odds of his happening are contingent on the player keeping the game running long enough to complete more than one run. It's also an aesthetic issue only, so there's no urgent need to fix.