Fixed theoretical incorrect trigger issue with the Mirage Arena cheevos
There was a theoretical flaw in the code that detects when an Arena Mission is actually completed, which is what drives those achievements to trigger in the first place. It checks to see if it equals the total number of rounds in the mission to determine if the player actually wins, but it didn't also check to see if the value had changed in the same frame. It's conceivable that this value could've had garbage data that wasn't cleared properly before the mission started.

Unfortunately, however, I was unable to reproduce the issue, and so this is unproven. It's possible the real culprit still lies in the code somewhere. But for now, this is the best I can do to resolve it.