Leaderboards and trigger update
Fixed an issue where leaderboards and their corresponding achievements would not keep time precisely the same way. Leaderboards are now considered a source of truth in regards to achievement times.

Also added Trigger flags to the speedrunning achievements to make more obvious in the future if there are more time discrepancies.